Private API

Note that the following methods are considered private methods, and changes in their behavior are not considered breaking changes.

BasicBSpline.r_nomial — Function

Calculate $r$-nomial coefficient

r_nomial(n, k, r)

\[(1+x+\cdots+x^r)^n = \sum_{k} a_{n,k,r} x^k\]

BasicBSpline._vec — Function

Convert AbstractKnotVector to AbstractVector

BasicBSpline._lower — Function

Internal methods for obtaining a B-spline space with one degree lower.

\[\begin{aligned} \mathcal{P}[p,k] &\mapsto \mathcal{P}[p-1,k] \\ D^r\mathcal{P}[p,k] &\mapsto D^{r-1}\mathcal{P}[p-1,k] \end{aligned}\]

Missing docstring.

Missing docstring for BasicBSpline._changebasis_R. Check Documenter's build log for details.

Missing docstring.

Missing docstring for BasicBSpline._changebasis_I. Check Documenter's build log for details.

BasicBSpline._changebasis_sim — Function

Return a coefficient matrix $A$ which satisfy

\[B_{(i,p_1,k_1)} = \sum_{j}A_{i,j}B_{(j,p_2,k_2)}\]

Assumption:

$P_1 ≃ P_2$

BasicBSplineFitting.innerproduct_R — Function

Calculate a matrix

\[A_{ij}=\int_{\mathbb{R}} B_{(i,p,k)}(t) B_{(j,p,k)}(t) dt\]

BasicBSplineFitting.innerproduct_I — Function

Calculate a matrix

\[A_{ij}=\int_{I} B_{(i,p,k)}(t) B_{(j,p,k)}(t) dt\]