Docstrings

All docstrings

BasicBSpline.BSplineManifold — Type

B-spline manifold for general polynomial degree

BasicBSpline.BSplineManifold — Method

convert AbstractBSplineManifold to BSplineManifold

BasicBSpline.BSplineSpace — Type

Construct B-spline space from given polynominal degree and knot vector.

\[\mathcal{P}[p,k]\]

BasicBSpline.BSplineSpace — Method

convert AbstractBSplineSpace to BSplineSpace

BasicBSpline.FastBSplineManifold — Type

B-spline manifold for lower polynomial degree TODO: make the field bsplinespaces to be conposite type, not abstract type, for performance

BasicBSpline.FastBSplineManifold — Method

convert AbstractBSplineManifold to FastBSplineManifold

BasicBSpline.FastBSplineSpace — Type

B-spline space for lower polynomial degree

BasicBSpline.FastBSplineSpace — Method

convert AbstractBSplineSpace to FastBSplineSpace

BasicBSpline.Knots — Type

Construct knot vector from given array.

\[k=(k_1,\dots,k_l)\]

BasicBSpline.bsplinebasis — Method

Multi-dimensional B-spline basis function.

\[B_{i^1,\dots,i^d}(t^1,\dots,t^d) =B_{(i^1,p^1,k^1)}(t^1)\cdots B_{(i^d,p^d,k^d)}(t^d)\]

BasicBSpline.bsplinebasis — Method

B-spline basis function. Modified version.

\[\begin{aligned} {B}_{(i,p,k)}(t) &= \frac{t-k_{i}}{k_{i+p}-k_{i}}{B}_{(i,p-1,k)}(t) +\frac{k_{i+p+1}-t}{k_{i+p+1}-k_{i+1}}{B}_{(i+1,p-1,k)}(t) \\ {B}_{(i,0,k)}(t) &= \begin{cases} &1\quad (k_{i}\le t<k_{i+1}<k_{l})\\ &1\quad (k_{i}\le t\le k_{i+1}=k_{l})\\ &0\quad (\text{otherwise}) \end{cases} \end{aligned}\]

BasicBSpline.bsplinebasis — Method

Multi-dimensional B-spline basis function.

\[B_{i^1,\dots,i^d}(t^1,\dots,t^d) =B_{(i^1,p^1,k^1)}(t^1)\cdots B_{(i^d,p^d,k^d)}(t^d)\]

BasicBSpline.bsplinebasis — Method

i-th B-spline basis function. Modified version.

BasicBSpline.bsplinebasis′ — Function

1st derivative of B-spline basis function. Modified version.

\[\dot{B}_{(i,p,k)}(t) =p\left(\frac{1}{k_{i+p}-k_{i}}B_{(i,p-1,k)}(t)-\frac{1}{k_{i+p+1}-k_{i+1}}B_{(i+1,p-1,k)}(t)\right)\]

BasicBSpline.bsplinebasis′ — Method

TODO: faster.....

BasicBSpline.bsplinebasis′₊₀ — Method

1st derivative of B-spline basis function. Right-sided limit version.

\[\dot{B}_{(i,p,k)}(t) =p\left(\frac{1}{k_{i+p}-k_{i}}B_{(i,p-1,k)}(t)-\frac{1}{k_{i+p+1}-k_{i+1}}B_{(i+1,p-1,k)}(t)\right)\]

BasicBSpline.bsplinebasis′₋₀ — Method

1st derivative of B-spline basis function. Left-sided limit version.

\[\dot{B}_{(i,p,k)}(t) =p\left(\frac{1}{k_{i+p}-k_{i}}B_{(i,p-1,k)}(t)-\frac{1}{k_{i+p+1}-k_{i+1}}B_{(i+1,p-1,k)}(t)\right)\]

BasicBSpline.bsplinebasis₊₀ — Method

B-spline basis function. Right-sided limit version.

\[\begin{aligned} {B}_{(i,p,k)}(t) &= \frac{t-k_{i}}{k_{i+p}-k_{i}}{B}_{(i,p-1,k)}(t) +\frac{k_{i+p+1}-t}{k_{i+p+1}-k_{i+1}}{B}_{(i+1,p-1,k)}(t) \\ {B}_{(i,0,k)}(t) &= \begin{cases} &1\quad (k_{i}\le t< k_{i+1})\\ &0\quad (\text{otherwise}) \end{cases} \end{aligned}\]

BasicBSpline.bsplinebasis₊₀ — Method

i-th B-spline basis function. Right-sided limit version.

BasicBSpline.bsplinebasis₋₀ — Method

B-spline basis function. Left-sided limit version.

\[\begin{aligned} {B}_{(i,p,k)}(t) &= \frac{t-k_{i}}{k_{i+p}-k_{i}}{B}_{(i,p-1,k)}(t) +\frac{k_{i+p+1}-t}{k_{i+p+1}-k_{i+1}}{B}_{(i+1,p-1,k)}(t) \\ {B}_{(i,0,k)}(t) &= \begin{cases} &1\quad (k_{i}< t\le k_{i+1})\\ &0\quad (\text{otherwise}) \end{cases} \end{aligned}\]

BasicBSpline.bsplinebasis₋₀ — Method

i-th B-spline basis function. Left-sided limit version.

BasicBSpline.bsplinesupport — Method

Return support of i-th B-spline basis function.

BasicBSpline.dim — Function

Calculate the dimension of B-spline manifold.

BasicBSpline.dim — Method

Return dimention of a B-spline space.

\[\dim(\mathcal{P}[p,k]) =\sharp k - p -1\]

BasicBSpline.fittingcontrolpoints — Method

Approximate given function by linear combination of B-spline functions. This function returns its control points. TODO: currently, this function only supports for 1-dim and 2-dim B-spline manifold.

BasicBSpline.mapping — Method

Calculate the mapping of B-spline manifold for given parameter.

\[\bm{p}(t^1,\dots,t^d) =\sum_{i^1,\dots,i^d}B_{i^1,\dots,i^d}(t^1,\dots,t^d) \bm{a}_{i^1,\dots,i^d}\]

BasicBSpline.mapping — Method

Calculate the mapping of B-spline manifold for given parameter.

\[\bm{p}(t^1,\dots,t^d) =\sum_{i^1,\dots,i^d}B_{i^1,\dots,i^d}(t^1,\dots,t^d) \bm{a}_{i^1,\dots,i^d}\]

BasicBSpline.refinement — Method

Refinement of B-spline manifold with given B-spline spaces.

BasicBSpline.refinement — Method

Refinement of B-spline manifold with additional degree and knots.

BasicBSpline.𝒫 — Method

Retrun FastBSplineSpace if ≤ MAX_DEGREE, or BSplineSpace if not.

\[\mathcal{P}[p,k]\]

Base.:⊆ — Method

Check inclusive relationship between B-spline spaces.

\[\mathcal{P}[p,k] \subseteq\mathcal{P}[p',k']\]

BasicBSpline.GaussianQuadrature_1dim — Method

fast, but only for 1-dim

BasicBSpline.GaussianQuadrature_2dim — Method

fast, but only for 2-dim

BasicBSpline.issqsubset — Method

Check inclusive relationship between B-spline spaces.

\[\mathcal{P}[p,k] \sqsubseteq\mathcal{P}[p',k'] \Leftrightarrow \mathcal{P}[p,k]|_{[k_{p+1},k_{l-p}]} \subseteq\mathcal{P}[p',k']|_{[\sharp k'_{p'+1},k'_{\sharp k'-p'}]}\]