Docstrings

All docstrings

BasicBSpline.BSplineManifold — Type

B-spline manifold for general polynomial degree

BasicBSpline.BSplineManifold — Method

Calculate the mapping of B-spline manifold for given parameter.

\[\bm{p}(t^1,\dots,t^d) =\sum_{i^1,\dots,i^d}B_{i^1,\dots,i^d}(t^1,\dots,t^d) \bm{a}_{i^1,\dots,i^d}\]

BasicBSpline.BSplineManifold — Method

convert AbstractBSplineManifold to BSplineManifold

BasicBSpline.BSplineSpace — Type

Construct B-spline space from given polynominal degree and knot vector.

\[\mathcal{P}[p,k]\]

This type BSplineSpace is slower than FastBSplineSpace, but this type is not limited with degree.

BasicBSpline.BSplineSpace — Method

convert AbstractBSplineSpace to BSplineSpace

BasicBSpline.FastBSplineManifold — Type

B-spline manifold for lower polynomial degree TODO: make the field bsplinespaces to be conposite type, not abstract type, for performance

BasicBSpline.FastBSplineManifold — Method

convert AbstractBSplineManifold to FastBSplineManifold

BasicBSpline.FastBSplineSpace — Type

B-spline space for lower polynomial degree

\[\mathcal{P}[p,k]\]

This type FastBSplineSpace is faster than BSplineSpace, but the degree must be equal or less than MAX_DEGREE.

BasicBSpline.FastBSplineSpace — Method

convert AbstractBSplineSpace to FastBSplineSpace

BasicBSpline.Knots — Type

Construct knot vector from given array.

\[k=(k_1,\dots,k_l)\]

Base.:⊆ — Method

Check inclusive relationship between B-spline spaces.

\[\mathcal{P}[p,k] \subseteq\mathcal{P}[p',k']\]

Base.issubset — Method

Check a inclusive relation ship $k\subset k'$.

\[(1,2,3) \subseteq (1,2,3,4) (1,2,5) \nsubseteq (1,2,3,4)\]

BasicBSpline._b_b_int — Method

Calculate

\[\int_{[k_{j+n-1}, k_{j+n}]} B_{(i,p,k)}(t) B_{(j,p,k)}(t) dt\]

Assumption:

$i ≤ j$

$1 ≤ n ≤ p-j+i+1$

BasicBSpline._b_b_int_I — Method

Calculate

\[\begin{aligned} &\int_{I} B_{(i,p,k)}(t) B_{(j,p,k)}(t) dt \\ \end{aligned}\]

BasicBSpline._b_b_int_R — Method

Calculate

\[\begin{aligned} &\int_{\mathbb{R}} B_{(i,p,k)}(t) B_{(j,p,k)}(t) dt \\ ={}& \begin{cases} \displaystyle \int_{[k_{j}, k_{i+p+1}]} B_{(i,p,k)}(t) B_{(j,p,k)}(t) dt & (i \le j) \\ \displaystyle \int_{[k_{i}, k_{j+p+1}]} B_{(i,p,k)}(t) B_{(j,p,k)}(t) dt & (j \le i) \end{cases} \end{aligned}\]

BasicBSpline._bsb0 — Method

Returns the value of $B_{(i,0,k)}(t)$. Assumption:

`$k_{i} ≤ t < k_{i+1}$

BasicBSpline._bsb1 — Method

Returns the values of $B_{(i,1,k)}(t), B_{(i+1,1,k)}(t)$. Assumption:

$k_{i} ≤ t < k_{i+1}$

BasicBSpline._bsb2 — Method

Returns the values of $B_{(i,2,k)}(t), B_{(i+1,2,k)}(t), B_{(i+2,2,k)}(t)$. Assumption:

$k_{i} ≤ t < k_{i+1}$

BasicBSpline._bsb3 — Method

Returns the values of $B_{(i,3,k)}(t), B_{(i+1,3,k)}(t), B_{(i+2,3,k)}(t), B_{(i+3,3,k)}(t)$. Assumption:

$k_{i} ≤ t < k_{i+1}$

BasicBSpline._bsb4 — Method

Returns the values of $B_{(i,4,k)}(t), B_{(i+1,4,k)}(t), B_{(i+2,4,k)}(t), B_{(i+3,4,k)}(t), B_{(i+4,4,k)}(t)$. Assumption:

$k_{i} ≤ t < k_{i+1}$

BasicBSpline._bsb5 — Method

Returns the values of $B_{(i,5,k)}(t), B_{(i+1,5,k)}(t), B_{(i+2,5,k)}(t), B_{(i+3,5,k)}(t), B_{(i+4,5,k)}(t), B_{(i+5,5,k)}(t)$. Assumption:

$k_{i} ≤ t < k_{i+1}$

BasicBSpline._changebasis_I — Method

Return a coefficient matrix A which satisfy

\[B_{(i,p,k)} = \sum_{j}A_{i,j}B_{(j,p',k')}\]

Assumption:

$P ⊑ P′$

BasicBSpline._changebasis_R — Method

Return a coefficient matrix A which satisfy

\[B_{(i,p,k)} = \sum_{j}A_{i,j}B_{(j,p',k')}\]

Assumption:

$P ⊆ P′$

BasicBSpline._changebasis_sim — Method

Return a coefficient matrix A which satisfy

\[B_{(i,p_1,k_1)} = \sum_{j}A_{i,j}B_{(j,p_2,k_2)}\]

Assumption:

$P1 ⊑ P2$

$P2 ⊑ P1$

BasicBSpline._knotindex — Method

Find an index $i$ such that $k_{i} ≤ t < k_{i+1}$ or $k_{i} < t = k_{i+1} = k_{ ext{end}}$.

BasicBSpline._knotindex₊₀ — Method

Find an index $i$ such that $k_{i} ≤ t < k_{i+1}$.

BasicBSpline._knotindex₋₀ — Method

Find an index $i$ such that $k_{i} < t ≤ k_{i+1}$.

BasicBSpline.bsplinebasis — Method

Multi-dimensional B-spline basis function.

\[B_{i^1,\dots,i^d}(t^1,\dots,t^d) =B_{(i^1,p^1,k^1)}(t^1)\cdots B_{(i^d,p^d,k^d)}(t^d)\]

BasicBSpline.bsplinebasis — Method

B-spline basis function. Modified version.

\[\begin{aligned} {B}_{(i,p,k)}(t) &= \frac{t-k_{i}}{k_{i+p}-k_{i}}{B}_{(i,p-1,k)}(t) +\frac{k_{i+p+1}-t}{k_{i+p+1}-k_{i+1}}{B}_{(i+1,p-1,k)}(t) \\ {B}_{(i,0,k)}(t) &= \begin{cases} &1\quad (k_{i}\le t<k_{i+1}<k_{l})\\ &1\quad (k_{i}\le t\le k_{i+1}=k_{l})\\ &0\quad (\text{otherwise}) \end{cases} \end{aligned}\]

BasicBSpline.bsplinebasis — Method

Multi-dimensional B-spline basis function.

\[B_{i^1,\dots,i^d}(t^1,\dots,t^d) =B_{(i^1,p^1,k^1)}(t^1)\cdots B_{(i^d,p^d,k^d)}(t^d)\]

BasicBSpline.bsplinebasis — Method

$i$-th B-spline basis function. Modified version.

BasicBSpline.bsplinebasis′ — Function

1st derivative of B-spline basis function. Modified version.

\[\dot{B}_{(i,p,k)}(t) =p\left(\frac{1}{k_{i+p}-k_{i}}B_{(i,p-1,k)}(t)-\frac{1}{k_{i+p+1}-k_{i+1}}B_{(i+1,p-1,k)}(t)\right)\]

BasicBSpline.bsplinebasis′₊₀ — Method

1st derivative of B-spline basis function. Right-sided limit version.

\[\dot{B}_{(i,p,k)}(t) =p\left(\frac{1}{k_{i+p}-k_{i}}B_{(i,p-1,k)}(t)-\frac{1}{k_{i+p+1}-k_{i+1}}B_{(i+1,p-1,k)}(t)\right)\]

BasicBSpline.bsplinebasis′₋₀ — Method

1st derivative of B-spline basis function. Left-sided limit version.

\[\dot{B}_{(i,p,k)}(t) =p\left(\frac{1}{k_{i+p}-k_{i}}B_{(i,p-1,k)}(t)-\frac{1}{k_{i+p+1}-k_{i+1}}B_{(i+1,p-1,k)}(t)\right)\]

BasicBSpline.bsplinebasis₊₀ — Method

B-spline basis function. Right-sided limit version.

\[\begin{aligned} {B}_{(i,p,k)}(t) &= \frac{t-k_{i}}{k_{i+p}-k_{i}}{B}_{(i,p-1,k)}(t) +\frac{k_{i+p+1}-t}{k_{i+p+1}-k_{i+1}}{B}_{(i+1,p-1,k)}(t) \\ {B}_{(i,0,k)}(t) &= \begin{cases} &1\quad (k_{i}\le t< k_{i+1})\\ &0\quad (\text{otherwise}) \end{cases} \end{aligned}\]

BasicBSpline.bsplinebasis₊₀ — Method

$i$-th B-spline basis function. Right-sided limit version.

BasicBSpline.bsplinebasis₋₀ — Method

B-spline basis function. Left-sided limit version.

\[\begin{aligned} {B}_{(i,p,k)}(t) &= \frac{t-k_{i}}{k_{i+p}-k_{i}}{B}_{(i,p-1,k)}(t) +\frac{k_{i+p+1}-t}{k_{i+p+1}-k_{i+1}}{B}_{(i+1,p-1,k)}(t) \\ {B}_{(i,0,k)}(t) &= \begin{cases} &1\quad (k_{i}< t\le k_{i+1})\\ &0\quad (\text{otherwise}) \end{cases} \end{aligned}\]

BasicBSpline.bsplinebasis₋₀ — Method

$i$-th B-spline basis function. Left-sided limit version.

BasicBSpline.bsplinesupport — Method

Return the support of $i$-th B-spline basis function.

\[\operatorname{supp}(B_{(i,p,k)})=[k_{i},k_{i+p+1}]\]

BasicBSpline.dim — Function

Calculate the dimension of B-spline manifold.

BasicBSpline.dim — Method

Return dimention of a B-spline space.

\[\dim(\mathcal{P}[p,k]) =\sharp k - p -1\]

BasicBSpline.fittingcontrolpoints — Method

Approximate given function by linear combination of B-spline functions. This function returns its control points.

func: Vector{<:Real} -> ℝ-vector space

BasicBSpline.fittingcontrolpoints — Method

func: (Real,Real,Real) -> ℝ-vector space

BasicBSpline.fittingcontrolpoints — Method

func: (Real,Real) -> ℝ-vector space

BasicBSpline.fittingcontrolpoints — Method

func: Real -> ℝ-vector space

BasicBSpline.integrate — Method

Integrate on rectangular solid

BasicBSpline.integrate — Method

Integrate on rectangular region.

BasicBSpline.integrate — Method

Integrate on interval

BasicBSpline.issqsubset — Method

Check inclusive relationship between B-spline spaces.

\[\mathcal{P}[p,k] \sqsubseteq\mathcal{P}[p',k'] \Leftrightarrow \mathcal{P}[p,k]|_{[k_{p+1},k_{l-p}]} \subseteq\mathcal{P}[p',k']|_{[\sharp k'_{p'+1},k'_{\sharp k'-p'}]}\]

BasicBSpline.refinement — Method

Refinement of B-spline manifold with given B-spline spaces.

BasicBSpline.refinement — Method

Refinement of B-spline manifold with additional degree and knots.

BasicBSpline.tangentvectors — Function

tangentvectors on B-spline manifold NOTICE: these methods are experimental